Refactored main.js into: windows, tray, camera, detection, and splash image logic
